Rename modules to match `ktor` plugins and migrate them to new ktor API
