+++ ktor-plugins/ktor-server-rsocket/src/commonMain/kotlin/io/rsocket/kotlin/ktor/server/Routing.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2015-2024 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.rsocket.kotlin.ktor.server
+
+import io.ktor.server.routing.*
+import io.ktor.server.websocket.*
+import io.rsocket.kotlin.*
+
+public fun Route.rSocket(protocol: String? = null, acceptor: ConnectionAcceptor) {
+    webSocket(protocol, rSocketHandler(acceptor))
+}
+
+public fun Route.rSocket(path: String, protocol: String? = null, acceptor: ConnectionAcceptor) {
+    webSocket(path, protocol, rSocketHandler(acceptor))
+}
