+++ ktor-plugins/rsocket-ktor-client/build.gradle.kts
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2015-2024 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import rsocketbuild.*
+
+plugins {
+    id("rsocketbuild.multiplatform-library")
+}
+
+description = "OLD ARTIFACT - migrate to ktor-client-rsocket"
+
+kotlin {
+    jvmTarget()
+    jsTarget()
+    nativeTargets()
+
+    sourceSets {
+        commonMain.dependencies {
+            implementation(projects.ktorClientRsocket)
+        }
+    }
+}
