--- rsocket-ktor/rsocket-ktor-server/api/rsocket-ktor-server.api
@@ -1,22 +1,0 @@
-public final class io/rsocket/kotlin/ktor/server/RSocketSupport {
-	public static final field Feature Lio/rsocket/kotlin/ktor/server/RSocketSupport$Feature;
-	public synthetic fun <init> (Lio/rsocket/kotlin/core/RSocketServer;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
-}
-
-public final class io/rsocket/kotlin/ktor/server/RSocketSupport$Config {
-	public final fun getServer ()Lio/rsocket/kotlin/core/RSocketServer;
-	public final fun server (Lkotlin/jvm/functions/Function1;)V
-	public final fun setServer (Lio/rsocket/kotlin/core/RSocketServer;)V
-}
-
-public final class io/rsocket/kotlin/ktor/server/RSocketSupport$Feature : io/ktor/server/application/BaseApplicationPlugin {
-	public fun getKey ()Lio/ktor/util/AttributeKey;
-	public fun install (Lio/ktor/server/application/Application;Lkotlin/jvm/functions/Function1;)Lio/rsocket/kotlin/ktor/server/RSocketSupport;
-	public synthetic fun install (Lio/ktor/util/pipeline/Pipeline;Lkotlin/jvm/functions/Function1;)Ljava/lang/Object;
-}
-
-public final class io/rsocket/kotlin/ktor/server/RoutingKt {
-	public static final fun rSocket (Lio/ktor/server/routing/Route;Ljava/lang/String;Ljava/lang/String;Lio/rsocket/kotlin/ConnectionAcceptor;)V
-	public static synthetic fun rSocket$default (Lio/ktor/server/routing/Route;Ljava/lang/String;Ljava/lang/String;Lio/rsocket/kotlin/ConnectionAcceptor;ILjava/lang/Object;)V
-}
-
