--- rsocket-ktor/src/commonMain/kotlin/io/rsocket/kotlin/ktor/dummy.kt
@@ -1,3 +1,0 @@
-package io.rsocket.kotlin.ktor
-
-private val dummy: Int = 0
